requirements.txt
git+https://github.com/Gawajn/segmentation-pytorch.git#egg=pixel-classifier-torch
numpy~=1.24.2
setuptools
matplotlib~=3.7.1
pandas~=1.5.3
seaborn
tqdm
albumentations~=1.3.0
scikit-image
opencv-python-headless
sklearn
future
scikit-learn
dataclasses~=0.6
scikit-learn
scipy~=1.9.1
mashumaro~=3.5
dataclasses_json
Pillow~=9.4.0
torch~=2.0.0
pylsd-nova